Update build-staging-deploy-s3.yml
Added copy on index.html so we can set headers for --cache-control including browser controls to not cache (max-age) and Cloud Front CDN controls to cache for 1 day (s-maxage). We want the browser to always check with Cloud Front for the lastest index.html with cache busting names baked into the js and css files. When new code is deployed to Cloud Front there is an invalidation run to clear Cloud Front.